pyproject.toml
[tool.poetry]
name = "kailio"
version = "0.1.0"
description = ""
authors = ["Andrew Kail <andrew.a.kail@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
flask = "^1.1.2"
flask-sqlalchemy = "^2.4.1"
Bootstrap-Flask = "^1.3.1"
Flask-Security-Too = "^3.4.2"
pytest-factoryboy = "^2.0.3"
pytest-pylint = "^0.17.0"
gunicorn = "^20.0.4"
Flask-Migrate = "^2.5.3"
Flask-Admin = "^1.5.6"
python-dotenv = "^0.13.0"
psycopg2 = "^2.8.5"
flask-ckeditor = "^0.4.4"
flask-shell-ipython = "^0.4.1"
flask-moment = "^0.10.0"
bcrypt = "^3.1.7"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
black = "^19.10b0"
pre-commit = "^2.4.0"
flake8 = "^3.8.1"
isort = "^4.3.21"
mypy = "^0.770"
pydocstyle = "^5.0.2"
coverage = "^5.1"
hypothesis = "^5.15.1"
factory-boy = "^2.12.0"
pytest-coverage = "^0.0"
safety = "^1.9.0"
factory_boy = "^2.12.0"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"

[tool.bandit]
pytest_filter = "*_test.py"
